Make UltiSnips work
Ended up using <Ctrl-space> instead of overloading tab for snippet
expansion/movement.  This is because there is no good way to
unambiguously tell if you want to use tab complete, or move to the next
snippets position.

Also ran into issues with custom snippet installation, and thus had to
extend the installation process to add some symbolic links.  All of the
linked files are ignored so that they can be upgraded by vundle if the
plugin is upgraded.